requirements.txt
agno
google-genai
google-search-results
python-dotenv
streamlit